fix: extract chartColors to eliminate lazy-load warning
Move COLORS, TOOL_COLORS, getToolColor from Charts.tsx to chartColors.ts
so RequestDashboard can import colors without pulling in Chart.js,
allowing LazyCharts dynamic import to code-split correctly.
